+++ src/supla/local_action.cpp
@@ -198,10 +198,10 @@ void LocalAction::disableAction(int32_t action,
   uint16_t eventToCheck = 0;
   uint16_t actionToCheck = 0;
   if (action >= 0 && action <= 65535) {
-    actionToCheck = action;
+    actionToCheck = static_cast<uint16_t>(action);
   }
   if (event >= 0 && event <= 65535) {
-    eventToCheck = event;
+    eventToCheck = static_cast<uint16_t>(event);
   }
 
   while (ptr) {
@@ -222,10 +222,10 @@ void LocalAction::enableAction(int32_t action,
   uint16_t eventToCheck = 0;
   uint16_t actionToCheck = 0;
   if (action >= 0 && action <= 65535) {
-    actionToCheck = action;
+    actionToCheck = static_cast<uint16_t>(action);
   }
   if (event >= 0 && event <= 65535) {
-    eventToCheck = event;
+    eventToCheck = static_cast<uint16_t>(event);
   }
   while (ptr) {
     if (ptr->trigger == this && (ptr->onEvent == eventToCheck || allEvents) &&
